+++ zcash_client_backend/src/data_api/testing/pool/dsl.rs
@@ -231,7 +231,7 @@ where
             .get_wallet_summary(confirmations_policy)
             .unwrap()?;
         let balance = binding.account_balances().get(&account.id())?;
-        Some(*balance)
+        Some(balance.clone())
     }
 
     /// Adds funds from a single note from an address of the given type.